Add a NamedAreaGridPlacement widget for placing in named regions
This new NamedAreaGridPlacement widget is also responsible for updating
the same GridParentData as the GridPlacement widget, but everything is
now name aware.

Notably, GridParentData has the notion of area name resolution. When the
data's area name changes, row and column data is cleared out, and awaits
populating via the placement algorithm.