Migration to friendlier names
CSS Grid Layout's property names are kind specific — if it has to do
with a template, as in, the aspects of a grid that are definite (not
related to automatic expansion), `grid-template-` precedes the property
name (like in the case of columns, rows, areas.)

But CSS has considerations we don't, like a single namespace for all
property names...and beyond that, I don't really like the `template-`
part of the `grid-template-` prefix in any case. The general case in
my mind is specifying the exact set of tracks required.

So I dropped "template" where from param/prop names.

Breaking change, obviously.