Add support for computeDrySize()
computeDrySize() is used to calculate the size of a RenderBox
without actually laying it out. Turns out the structure of the
RenderLayoutGrid was already well suited to this task, as our sizing
calculations are independent from layout.

I also added a couple of tests to make sure that we satisfy the
requirements of the function — that the grid's "dry" size is the same as
the "wet" size, and that the `layout()` methods of the grid's children
are never called in this instance.

Fixes #42