Improve layout speed of periodic table
The periodic table's performance was being heavily impacted by text
layout.

By wrapping the `AtomicElementWidget`s in `AspectRatio` widgets (and
expecting a touch of overflow), performance has been drastically
improved.

We're still impacted by text layout performance, but to a much lesser
degree. This could probably be resolved by scaling the table through
a Transform, but for now this is a big improvement.

Fixes #33